requirements.txt
selenium
pytest
allure-pytest
pydantic
ruff
requests
python-dotenv
pyyaml
presidio-analyzer
spacy
webdriver-manager
